refactor news adapter
